Actualización backend ordenes

+++ backend/src/main/java/com/libruarys/repository/UsuarioRepository.java
@@ -4,6 +4,6 @@ import com.libruarys.model.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {  // ← Cambiar Long a Integer
     Optional<Usuario> findByCorreo(String correo);
 }
